Patch decode_involved_dimensions to avoid skipping EP collectives
This patch avoids EP collectives to be ignored in case TP = 1; we just pass the max of TP and EP. The patch works because this dimension splitting was part of astra-sim but SimAI does not seem to really use it; in particular, MockNccl does not use the resulting topology to determine the communication flows (it uses MockNcclGroup::getFlowModel instead); however NcclTreeFlowModel uses it to get the number of nodes in the ring. To properly fix this there are two options: (i) remove this whole process and uniform the simulator to only use the output of getFlowModel; (ii) generalize this to also work for the case of overlapped splits (EP and TP) and uniform the SimAI code to use this output instead of generating a new one from getFlowModel.